+++ client/src/game/shapes/shape.ts
@@ -264,6 +264,29 @@ export abstract class Shape {
         }
     }
 
+    drawSelection(ctx: CanvasRenderingContext2D): void {
+        ctx.globalCompositeOperation = this.globalCompositeOperation;
+        const z = gameStore.zoomFactor;
+        const bb = this.getBoundingBox();
+        ctx.strokeRect(g2lx(bb.topLeft.x), g2ly(bb.topLeft.y), bb.w * z, bb.h * z);
+
+        // Draw vertices
+        for (const p of this.points) {
+            ctx.beginPath();
+            ctx.arc(g2lx(p[0]), g2ly(p[1]), 3, 0, 2 * Math.PI);
+            ctx.fill();
+        }
+
+        // Draw edges
+        ctx.beginPath();
+        ctx.moveTo(g2lx(this.points[0][0]), g2ly(this.points[0][1]));
+        for (let i = 1; i <= this.points.length; i++) {
+            const vertex = this.points[i % this.points.length];
+            ctx.lineTo(g2lx(vertex[0]), g2ly(vertex[1]));
+        }
+        ctx.stroke();
+    }
+
     getInitiativeRepr(): InitiativeData {
         return {
             uuid: this.uuid,