+++ CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to this project will be documented in this file.
 - Pasted shapes are now pasted relative to the screen position
 - [tech] Mousemove events are now throttled, so that they don't fire a gazillion events.
 - [tech] tslint swapped out for eslint
+- [tech] Refactor Layer.draw to use Shape.drawSelection
 
 ### Fixed